+++ src/firmware/lib/blinkscript/src/vm/model/string_hash.hpp
@@ -9,7 +9,7 @@ namespace SyncBlink
     inline uint32_t stringHash(const std::string& value)
     {
         uint32_t hash = 2166136261u;
-        for (int i = 0; i < value.length(); i++) {
+        for (size_t i = 0; i < value.length(); i++) {
             hash ^= (uint8_t)value[i];
             hash *= 16777619;
         }
